+++ src/cryptography/hazmat/backends/openssl/dsa.py
@@ -225,7 +225,6 @@ class _DSAPublicKey(object):
         return self._backend._public_key_bytes(
             encoding,
             format,
-            None,
             evp_pkey,
             None
         )