+++ src/cryptography/hazmat/backends/openssl/rsa.py
@@ -609,7 +609,6 @@ class _RSAPublicKey(object):
         return self._backend._public_key_bytes(
             encoding,
             format,
-            self._backend._lib.PEM_write_bio_RSAPublicKey,
             self._evp_pkey,
             self._rsa_cdata
         )